Add Shape inference for plugins
