+++ tensorflow/compiler/tf2tensorrt/utils/trt_plugin_shape_function.h
@@ -1,0 +1,27 @@
+/* Copyright 2016 The TensorFlow Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+#ifndef TENSORFLOW_COMPILER_TF2TENSORRT_UTILS_TRT_PLUGIN_SHAPE_FUNCTION_H_
+#define TENSORFLOW_COMPILER_TF2TENSORRT_UTILS_TRT_PLUGIN_SHAPE_FUNCTION_H_
+
+#include "tensorflow/core/framework/shape_inference.h"
+#include "tensorflow/compiler/tf2tensorrt/convert/utils.h"
+namespace tensorflow {
+namespace tensorrt {
+namespace shape_inference {
+Status TRTPluginShapeFunction(tensorflow::shape_inference::InferenceContext* c);
+}
+}  // namespace tensorrt
+}  // namespace tensorflow
+#endif
